v4.2.0 - 21.02.2019:
For BRFv5 we reimplemented Optical Flow. This single component does now replace the BRFv4 version of it.

Since macOS Mojave it's not possible to run a camera app in the Xcode debug terminal.
Xcode is not in the list of allowed apps to use the camera. So to be able to run a terminal camera app in mac, I had to edit the Scheme to run on the system Terminal.app